Tests/LibWeb/Text/expected/DOM/Element-getAttributeNS.txt

xlink:ref getAttribute = 'null'
xlink:ref getAttributeNS = 'test'
href getAttribute = 'test'
href getAttributeNS = 'null'
